requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
python-multipart>=0.0.9
pydantic>=2.8,<3.0
pysam>=0.23,<1.0
openpyxl>=3.1,<4.0
Pillow>=10.0,<12.0
nibabel>=5.2,<6.0
